+++ Common/Footsteps/NpcFootsteps.cs
@@ -4,10 +4,7 @@
 
 using System.Linq;
 using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using Terraria;
-using Terraria.Audio;
-using Terraria.GameContent;
 using Terraria.ID;
 using Terraria.ModLoader;
 using TerrariaOverhaul.Core.Configuration;
@@ -23,11 +20,15 @@ internal sealed class NpcFootsteps : GlobalNPC
 		public Rectangle[] FootstepFrames = [];
 		public FootstepSounds? SoundsOverride;
 		public Vector2 OldVelocity;
+		public Vector2 OldPosDelta;
+		public Vector2 OldPosition;
 	}
 
 	private const uint FootstepCooldown = 3;
 
-	public static readonly ConfigEntry<bool> EnableNpcFootsteps = new(ConfigSide.ClientOnly, true, "Ambience", "Enemies");
+	public static readonly ConfigEntry<bool> EnableNpcFootstepSounds = new(ConfigSide.ClientOnly, true, "Ambience", "Enemies");
+	public static readonly ConfigEntry<bool> EnableNpcFootstepInteractions = new(ConfigSide.ClientOnly, true, "Enemies", "BloodAndGore");
+	public static readonly ConfigEntry<bool> EnableNpcLandingScreenShake = new(ConfigSide.ClientOnly, true, "Enemies", "Camera");
 	private static readonly Rectangle[] genericWithIdle = [new(0, 5 * 56, 40, 56), new(0, 12 * 56, 40, 56)];
 	private static readonly Rectangle[] genericWithoutIdle = [new(0, 4 * 56, 40, 56), new(0, 11 * 56, 40, 56)];
 
@@ -65,9 +66,8 @@ internal sealed class NpcFootsteps : GlobalNPC
 
 	private void UpdateFootsteps(NPC npc)
 	{
-		if (Main.dedServ || !EnableNpcFootsteps || Data is not FootstepData data) {
-			return;
-		}
+		if (Main.dedServ || Data is not FootstepData data) return;
+		if (!EnableNpcFootstepSounds && !EnableNpcFootstepInteractions && !EnableNpcLandingScreenShake) return;
 
 		var tick = Main.GameUpdateCount;
 		var onGround = npc.velocity.Y == 0f;
@@ -86,17 +86,25 @@ internal sealed class NpcFootsteps : GlobalNPC
 			footstepType = FootstepType.Default;
 		}
 
+		var posDelta = Data.OldPosition != default ? (npc.position - data.OldPosition) : default;
+		if (npc.boss) posDelta *= 2;
+
 		if (footstepType != null && (tick - data.LastFootstepTick) > FootstepCooldown && FootstepSystem.Footstep(new() {
 			Kind = footstepType.Value,
 			Hitbox = npc.Hitbox,
-			Velocity = (npc.velocity, npc.boss ? data.OldVelocity * 2f : npc.velocity),
-			Volume = 0.5f,
+			Velocity = (posDelta, data.OldPosDelta),
 			SoundsOverride = data.SoundsOverride,
+			Volume = EnableNpcFootstepSounds ? 0.5f : 0,
+			AllowGoreInteraction = EnableNpcFootstepInteractions,
+			AllowScreenShake = EnableNpcLandingScreenShake,
+			AllowParticles = !npc.wet,
 		})) {
 			data.LastFootstepTick = tick;
 		}
 
 		data.PreviousFrame = npc.frame;
 		data.OldVelocity = npc.velocity;
+		data.OldPosition = npc.position;
+		data.OldPosDelta = posDelta;
 	}
 }